% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbet.R
\name{cbet}
\alias{cbet}
\title{Computes Counting Betweenness}
\usage{
cbet(A)
}
\arguments{
\item{A}{The adjacency matrix of the network to be analyzed. It must be square.}
}
\value{
The vector containing the values of Counting Betweenness of the network..
}
\description{
Counting Betweenness implemented as in DePaolis et al(2022)
}
\examples{
cbet(exmpl_matrix)
}
