% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbet_norm.R
\name{cbet_norm}
\alias{cbet_norm}
\title{Computes Counting Betweenness in normalized format}
\usage{
cbet_norm(A)
}
\arguments{
\item{A}{The adjacency matrix of the network to be analyzed.It must be square.}
}
\value{
The vector containing the normalized values (between 0 and 1) of Counting Betweenness of the network.
}
\description{
A normalized version of Counting Betweenness implemented as in DePaolis et al(2022)
}
\examples{
cbet_norm(exmpl_matrix)
}
