% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exmpl_matrix}
\alias{exmpl_matrix}
\title{Data to showcase the functions in the xtranat package}
\format{
A 10 by 10 square matrix
}
\source{
{Created by the author as an example}
}
\usage{
exmpl_matrix
}
\description{
Contains a randomly created adjacency matrix
}
\details{
It is a 10 by 10 matrix with some values in the diagonal to represent loops
}
\examples{
data(exmpl_matrix)
}
\keyword{datasets}
