% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfpt.R
\name{mfpt}
\alias{mfpt}
\title{Computes mfpt}
\usage{
mfpt(A)
}
\arguments{
\item{A}{The adjacency matrix of the network to be analyzed}
}
\value{
mfpt.
}
\description{
mean first-passage time implemented as in DePaolis et al(2022)
}
