% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwc.R
\name{rwc}
\alias{rwc}
\title{Computes Random Walk Centrality}
\usage{
rwc(A)
}
\arguments{
\item{A}{The adjacency matrix of the network to be analyzed.It must be square.}
}
\value{
The vector containing the values of Random Walk Centrality of the network.
}
\description{
Random Walk Centrality implemented as in DePaolis et al(2022)
}
\examples{
rwc(exmpl_matrix)
}
