% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwc_norm.R
\name{rwc_norm}
\alias{rwc_norm}
\title{Computes Random Walk Centrality in normalized format}
\usage{
rwc_norm(A)
}
\arguments{
\item{A}{The adjacency matrix of the network to be analyzed.It must be square.}
}
\value{
The vector containing the normalized values (between 0 and 1) of Random Walk Centrality of the network.
}
\description{
A normalized version of Random Walk Centrality implemented as in DePaolis et al(2022)
}
\examples{
rwc_norm(exmpl_matrix)
}
